Refactor Home page with RouterGuard
